Symlink Google Drive Data Space to Local Data Directory

This function creates a symbolic link from a Google Drive data space (specified in a config file) to a local "data" directory. It checks for the existence of the Google Drive path and the local data directory, creates the local directory if it does not exist, and then creates the symlink.

Usage

link_inputs(cfg_path)

Arguments

cfg_path

Character. Path to the configuration file containing the Google Drive data space path.

Value

Logical. Returns TRUE if the linking is successful; otherwise, stops with an error.

Examples

if (FALSE) { # \dontrun{
link_inputs("config.yml")
} # }

link_inputs("config.yml")
#> Linking the google drive data space to the local Data directory: /n/holylabs/LABS/cgolden_lab/Lab/data_freeze/golden_googledrive_rclone/Climate-Smart Public Health - Madagascar/4. Datasets/DHS Data
#> Warning: cannot symlink '/n/holylabs/LABS/cgolden_lab/Lab/data_freeze/golden_googledrive_rclone/Climate-Smart Public Health - Madagascar/4. Datasets/DHS Data' to '/n/holylabs/cgolden_lab/Lab/projects/DHSHarmonization/data/DHS Data', reason 'File exists'
#> Linking successful!
#> [1] TRUE